Added base_adapter.
